backend/app/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiopg==1.4.0
aiosignal==1.3.2
annotated-types==0.7.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
bcrypt==4.1.3
bitarray==3.0.0
certifi==2024.8.30
cffi==1.16.0
channels==4.0.0
charset-normalizer==3.4.0
ckzg==2.0.1
constantly==23.10.4
cryptography==42.0.8
cytoolz==1.0.1
daphne==4.1.2
Django==4.2.13
django-annoying==0.10.6
django-cors-headers==4.4.0
django-environ==0.11.2
djangochannelsrestframework==1.2.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
eth-account==0.13.4
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.6.0
eth-rlp==2.1.0
eth-typing==5.0.1
eth-utils==5.1.0
eth_abi==5.1.0
frozenlist==1.5.0
hexbytes==1.2.1
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
json-log-formatter==0.5.1
Markdown==3.6
msgpack==1.0.8
multidict==6.1.0
parsimonious==0.10.0
pillow==10.4.0
propcache==0.2.1
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.3
pydantic_core==2.27.1
PyJWT==2.8.0
pyOpenSSL==24.1.0
pyotp==2.9.0
pyunormalize==16.0.0
qrcode==8.0
regex==2024.11.6
requests==2.32.3
rlp==4.0.1
service-identity==24.1.0
six==1.16.0
sqlparse==0.5.0
toolz==1.0.0
Twisted==24.3.0
txaio==23.1.1
types-requests==2.32.0.20241016
typing_extensions==4.12.2
urllib3==2.2.3
uuid==1.30
web3==7.6.0
websockets==13.1
yarl==1.18.3
zope.interface==6.4.post2
